fix: optionally show gauges in PoolSelector, moves validation params to constants
